+++ src/packages/user/user-group/collection/manifests.ts
@@ -1,10 +1,8 @@
-import {
-	UMB_USER_GROUP_COLLECTION_REPOSITORY_ALIAS,
-	manifests as repositoryManifests,
-} from './repository/manifests.js';
+import { manifests as repositoryManifests } from './repository/manifests.js';
 import { manifests as collectionViewManifests } from './views/manifests.js';
 import { manifests as collectionActionManifests } from './action/manifests.js';
 import { UMB_USER_GROUP_COLLECTION_ALIAS } from './constants.js';
+import { UMB_USER_GROUP_COLLECTION_REPOSITORY_ALIAS } from './repository/constants.js';
 import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
 
 const collection: ManifestTypes = {
