+++ src/packages/user/user-group/collection/repository/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_USER_GROUP_COLLECTION_REPOSITORY_ALIAS = 'Umb.Repository.UserGroupCollection';
