+++ src/packages/user/user-group/repository/detail/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_USER_GROUP_DETAIL_REPOSITORY_ALIAS = 'Umb.Repository.UserGroup.Detail';
+export const UMB_USER_GROUP_DETAIL_STORE_ALIAS = 'Umb.Store.UserGroup.Detail';
