+++ src/packages/user/user-group/repository/detail/index.ts
@@ -1,5 +1,5 @@
 export { UmbUserGroupDetailRepository } from './user-group-detail.repository.js';
-export { UMB_USER_GROUP_DETAIL_REPOSITORY_ALIAS } from './manifests.js';
-export { UMB_USER_GROUP_DETAIL_STORE_CONTEXT } from './user-group-detail.store.js';
+export { UMB_USER_GROUP_DETAIL_REPOSITORY_ALIAS } from './constants.js';
+export { UMB_USER_GROUP_DETAIL_STORE_CONTEXT } from './user-group-detail.store.token.js';
 
 export type { UmbUserGroupDetailStore } from './user-group-detail.store.js';
