+++ src/packages/user/user-group/repository/detail/user-group-detail.repository.ts
@@ -1,6 +1,6 @@
 import type { UmbUserGroupDetailModel } from '../../types.js';
 import { UmbUserGroupServerDataSource } from './user-group-detail.server.data-source.js';
-import { UMB_USER_GROUP_DETAIL_STORE_CONTEXT } from './user-group-detail.store.js';
+import { UMB_USER_GROUP_DETAIL_STORE_CONTEXT } from './user-group-detail.store.token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbDetailRepositoryBase } from '@umbraco-cms/backoffice/repository';
