+++ src/packages/user/user-group/repository/item/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_USER_GROUP_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.UserGroupItem';
+export const UMB_USER_GROUP_STORE_ALIAS = 'Umb.Store.UserGroupItem';
