+++ src/packages/user/user-group/repository/item/index.ts
@@ -1,3 +1,3 @@
 export { UmbUserGroupItemRepository } from './user-group-item.repository.js';
-export { UMB_USER_GROUP_ITEM_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_USER_GROUP_ITEM_REPOSITORY_ALIAS } from './constants.js';
 export type { UmbUserGroupItemModel } from './types.js';
